remove legacy and duplicated CMakeDeps tests
